fix(runs): surface stalled burrow stream in the UI (warren-6376)
When burrow's unix socket times out under CPU starvation the stream
bridge reconnects with exponential backoff but nothing surfaced to the
user — runs sat at "starting" indefinitely.

- runWithReconnect now emits a one-shot `bridge_stalled` (stream:system)
  event after BRIDGE_STALL_THRESHOLD=3 consecutive errored reconnects
  with no forward progress, and a matching `bridge_recovered` once events
  stream again (forward progress resets the counter and clears the stall).
- RunDetail.tsx renders an "Agent infrastructure unreachable" warning
  banner while a non-terminal run is stalled (isBridgeStalled scans the
  events tail; banner clears on recovery).

To stay under the per-file size ratchet, extracted the reconnect/stall
machinery into src/server/bridge-reconnect.ts and shared bridge test
fixtures into src/server/bridges.test-helpers.ts. Trimmed dev tracker/SPEC
refs from user-facing title= tooltips in RunDetail.tsx to keep the raw-JS
bundle under its (tight, temporary pl-55a3) ceiling.
