+++ .mulch/expertise/build.jsonl
@@ -91,3 +91,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T22:16:53.909Z","evidence":{"seeds":"warren-d0a9","commit":"b0c91170cc71f3a7fea42c3d0eef4bdb2912e145"},"tags":["refactor","file-size-ratchet","pl-9088"],"relates_to":["mx-c7ca2b"],"dir_anchors":["src/preview"],"content":"Preview eviction split (warren-d0a9 / pl-9088 step 8): src/preview/eviction.ts (819 lines) → src/preview/eviction/{config,types,tick,repo,sidecar,worker}.ts; eviction.test.ts (545) → {config,tick,worker}.test.ts + test-helpers.ts. index.ts re-exports the public surface so call sites (server/main.ts, handlers/{diagnostics,runs}.ts, diagnostics/checks.ts, preview/teardown.ts) only change the suffix to /index.ts. Dropped the file-size budget entry and the noExcessiveCognitiveComplexity grandfather entry for the old monolith; new complexity hot spots refactored via helper extraction (tryLoadWarrenConfig+parseOverrideDuration + stopSidecarsForRow+deleteSidecarSafely+emitEvictedEvent in tick; lockRowForPostgres+readCurrentPreview+writeTornDown+shapeNonClaimable in repo; buildTickInput+runTickAndCount in worker).","id":"mx-5c4f4d"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T22:31:09.896Z","evidence":{"seeds":"warren-62a7","commit":"c18399e38f556ec6e26324a5ff27941a8812b227"},"tags":["warren-62a7","pl-9088","refactor"],"relates_to":["mx-5c4f4d"],"dir_anchors":["src/runs/reap"],"content":"Preview launch split (warren-62a7 / pl-9088 step 9): src/preview/launch.ts (712 lines) → src/preview/launch/{index,types,probe,helpers,setup,orchestrate,url}.ts. Test split: orchestrate.test.ts (main launchPreview cases), probe-loop.test.ts (warren-9b15 two-phase), setup.test.ts (warren-d9e7), probe.test.ts (tcpConnectOnce direct), url.test.ts. Shared fakes in test-helpers.ts (FakeSidecar, fakeSidecars, fakeFetch, fakeTcp, alwaysConnected/Refused, PREVIEW_CONFIG, setupLaunchEnv). Public API re-exported from launch/index.ts — callers just suffix '/index.ts'. file-size-budgets dropped both old entries; biome cognitive-complexity override moved to orchestrate.ts + setup.ts (both ~16); lines-per-function override removed entirely.","id":"mx-b4b3e1"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-28T19:12:47.143Z","evidence":{"seeds":"warren-23fe","commit":"a41bec751b93d79d943519533a05328ac8688990"},"relates_to":["mx-3c5052"],"content":"Bundle-size ratchet (warren-5abc) has a documented escape hatch for foundational growth: when a Phase legitimately expands the surface (e.g. warren-23fe Phase 1 design tokens added the status palette, radius/shadow/z scales, brand color, and self-hosted font CSS), raise the floors in scripts/bundle-size-budgets.json with an inline $comment_<bucket>_<date> field naming the seed + plan step + what was added + what was trimmed to minimise the bump + the named child seeds that must hold flat against the new floor. This mirrors the file-size grandfather pattern: a one-time legitimate raise with a tracker rationale, not a silent ratchet weakening. Don't use this for casual additions or dep bumps — those still owe a code-split or trim. For warren-23fe CSS went raw 34474->35200 (+726B) / gzip 6960->7160 (+200B); we trimmed unused -foreground pairs (success/warning/info/danger/neutral/card/accent), --shadow-xl, and --z-base to keep the bump minimal.","id":"mx-f1f5f7"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-29T02:13:34.165Z","evidence":{"seeds":"warren-6376","commit":"2a0483bc662b04aae9a485189af4cbec5ab7ea2a"},"content":"src/server/bridge-reconnect.ts (extracted from bridges.ts to stay under the file-size ratchet) holds runWithReconnect: after BRIDGE_STALL_THRESHOLD=3 consecutive errored reconnects with no forward progress it emits a one-shot bridge_stalled (stream:system) event, and bridge_recovered once events stream again (result.written>0 resets the counter + clears the stall). UI: RunDetail.tsx isBridgeStalled() scans the events tail and renders a warning Alert banner while !isTerminal. Shared bridge test fixtures now live in src/server/bridges.test-helpers.ts. Gotcha: the raw-JS bundle budget (584000) sits at its temporary pl-55a3 ceiling with ~0 headroom, so the banner only fit after trimming dev tracker/SPEC refs from user-facing title= tooltips.","id":"mx-9eb278"}
