+++ biome.json
@@ -105,6 +105,7 @@
 				"src/runs/stream/bridge.ts",
 				"src/runs/usage-aggregate.ts",
 				"src/server-config/workers.ts",
+				"src/server/bridge-reconnect.ts",
 				"src/server/bridges.ts",
 				"src/server/errors.ts",
 				"src/server/handlers/plan-runs.ts",
