+++ scripts/file-size-budgets.json
@@ -18,8 +18,7 @@
 		"src/runs/pause.test.ts": 616,
 		"src/runs/pr.test.ts": 604,
 		"src/runs/pr.ts": 659,
-		"src/server/bridges.test.ts": 647,
-		"src/server/bridges.ts": 655,
+		"src/server/bridges.test.ts": 557,
 		"src/server/handlers/agents.test.ts": 533,
 		"src/server/handlers/projects.settings.test.ts": 685,
 		"src/server/handlers/plot-plan-runs.test.ts": 640,
